Move plotting utility functions to new utils.plot
